Move to alpine and improve security & smoke test
