--- integrations/_infra/init.sh
@@ -1,5 +1,0 @@
-if test -e /usr/local/share/ca-certificates/cert.crt; then
-  update-ca-certificates
-fi
-
-ocean sail
