+++ port_ocean/tests/test_smoke.py
@@ -4,7 +4,7 @@ import pytest
 
 from port_ocean.clients.port.client import PortClient
 from port_ocean.clients.port.types import UserAgentType
-from port_ocean.tests.helpers.fixtures import SmokeTestDetails
+from port_ocean.tests.helpers.smoke_test import SmokeTestDetails
 
 
 pytestmark = pytest.mark.smoke
